# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reunite API
  version: 1.0.0
  description: Lost-and-found recovery network. Public responses are teasers only.
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local
...
paths:
...
  /reports:
    get:
      tags: [Reports]
      summary: Search teasers
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [lost, found] }
        - in: query
          name: category_id
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: Paginated teasers }
    post:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      summary: Create draft report
      responses:
        '201': { description: Created }
...
  /claims:
    post:
      tags: [Claims]
      security: [{ bearerAuth: [] }]
      summary: Start a claim
      responses:
        '201': { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...